Add ingest_metadata program to ingest cocreator metadata into block metadata ready to publish.
Add missing buy_mystery transaction to nifty_stakes.js.

Include maximum_price_lamports in buy and buy_mystery tx so that a user is
never surprised if a faulty UI shows a lower than actual price.

Added refresh_tokens and refresh_stakes to nifty_stakes.js so that after
an action that would change wallet contents, a UI update can be immediate.

Fixed nifty_stakes.js Block and Entry refresh.

Sign the metaplex metadata with the Program Authority key so that its creator
entry can be Verified, which will make it simpler and easier for users to
verify Shinobi NFTs as being authentic.

Fixed a bug in reading metaplex metadata from on-chain data which was causing
creators to be lost on setting level metadata.

Update metadata publishing script:
  - No more full_frame_N.png
  - card_N.png becomes image_N.png
  - preview_N.jpg and thumb_N.jpg are added
  - max_ki_factor is taken out of config, ki_factor is completely controlled
    by the ingest_metadata script
  - Use aws sync instead of aws cp so that files are not constantly
    overwritten when they haven't changed.  Because the filename is the
    SHA-256, can just rely on file sizes, not contents, for detecting changed
    files
  - Properly set content-type on AWS files
  - Added in publishing of entry art details, like which hand, face,
    background, etc, it used

Work around an apparent issue with valid_block_height as reported by RPC nodes,
it's possible that they are hardcoded with 0.4 second slot times or something.